# Simple macro benchmark