@zregvart 2024-12-03
I ran the benchmark on main (this branch) and with reverted #2184, this gives us some data to work with:
```
$ make benchmark > new.txt
$ git revert --no-commit --mainline 1 b8c77a098e2e0aa661d2f61cb7d44569494fe195
$ make benchmark > old.txt
$ go run golang.org/x/perf/cmd/benchstat@latest old.txt new.txt 
pkg: golang.org/x/benchmarks
goos: linux
goarch: amd64
          │   old.txt   │             new.txt             │
          │   sec/op    │   sec/op     vs base            │
Simple-12   21.08 ± 24%   22.64 ± 28%  ~ (p=0.475 n=10+7)

          │       old.txt        │                 new.txt                  │
          │ GC-bytes-from-system │ GC-bytes-from-system  vs base            │
Simple-12           7.450Mi ± 4%           7.452Mi ± 0%  ~ (p=1.000 n=10+7)

          │   old.txt   │             new.txt              │
          │ STW-sec/GC  │  STW-sec/GC   vs base            │
Simple-12   72.11µ ± 5%   76.71µ ± 15%  ~ (p=0.109 n=10+7)

          │   old.txt   │             new.txt              │
          │ STW-sec/op  │  STW-sec/op   vs base            │
Simple-12   3.678m ± 8%   3.989m ± 10%  ~ (p=0.193 n=10+7)

          │      old.txt       │                new.txt                 │
          │ allocated-bytes/op │ allocated-bytes/op  vs base            │
Simple-12         3.531Gi ± 0%         3.529Gi ± 0%  ~ (p=0.601 n=10+7)

          │   old.txt   │             new.txt             │
          │  allocs/op  │  allocs/op   vs base            │
Simple-12   74.91M ± 0%   74.90M ± 0%  ~ (p=0.417 n=10+7)

          │      old.txt      │                new.txt                │
          │ bytes-from-system │ bytes-from-system  vs base            │
Simple-12        336.5Mi ± 6%        331.6Mi ± 8%  ~ (p=0.109 n=10+7)

          │        old.txt         │                  new.txt                   │
          │ heap-bytes-from-system │ heap-bytes-from-system  vs base            │
Simple-12             313.2Mi ± 5%             309.3Mi ± 7%  ~ (p=0.103 n=10+7)

          │         old.txt         │                   new.txt                   │
          │ other-bytes-from-system │ other-bytes-from-system  vs base            │
Simple-12             13.06Mi ± 22%             12.19Mi ± 18%  ~ (p=0.315 n=10+7)

          │    old.txt     │              new.txt               │
          │ peak-RSS-bytes │ peak-RSS-bytes  vs base            │
Simple-12     374.0Mi ± 4%     375.0Mi ± 4%  ~ (p=0.216 n=10+7)

          │    old.txt    │              new.txt              │
          │ peak-VM-bytes │ peak-VM-bytes  vs base            │
Simple-12    2.594Gi ± 0%    2.593Gi ± 0%  ~ (p=0.364 n=10+7)

          │         old.txt         │                   new.txt                   │
          │ stack-bytes-from-system │ stack-bytes-from-system  vs base            │
Simple-12             2.531Mi ± 15%             2.625Mi ± 32%  ~ (p=0.757 n=10+7)

          │     old.txt     │                 new.txt                  │
          │ user+sys-sec/op │ user+sys-sec/op  vs base                 │
Simple-12        16.81 ± 1%        17.29 ± 5%  +2.89% (p=0.003 n=10+7)
```