@simonbaird 2025-02-04
I'll give it a quick rebase to fix the go.mod conflict, then I suggest merging. Lgtm.