@simonbaird 2025-02-04
I tried to run it locally and hit this:

```
ERRO[0013] new digest: a digest must contain exactly one '@' separator (e.g. registry/repository@digest) saw: registry.access.redhat.com/ubi9/ubi-micro@sha256:111395a0b54d220c4ad821602a4a28807d6c4fb2feb19b43a6468a9d5f257463@sha256:7ebc389c674b94d5604f7ababe7f026e2505b4664ce1af6caf0f38c27a09ccbe  ref="registry.access.redhat.com/ubi9/ubi-micro@sha256:111395a0b54d220c4ad821602a4a28807d6c4fb2feb19b43a6468a9d5f257463@sha256:7ebc389c674b94d5604f7ababe7f026e2505b4664ce1af6caf0f38c27a09ccbe" rego=ec.oci.image_manifest
```

which I think is the same problem as https://issues.redhat.com/browse/EC-1041 .