@simonbaird 2025-02-07
I'm hitting a different problem now trying to run this locally, so not confident about merging.

Because I'm adding extra commits, I'll redo the PR with my fork, see #2304.